Moved Navbar to Shade
