+++ apps/shade/src/index.ts
@@ -11,6 +11,7 @@ export * from './components/ui/dropdown-menu';
 export * from './components/ui/form';
 export * from './components/ui/input';
 export * from './components/ui/label';
+export * from './components/ui/navbar';
 export * from './components/ui/popover';
 export * from './components/ui/right-sidebar';
 export * from './components/ui/separator';
